Modified the form_button_helper to display a unique set of nav options to users who are editing their profile

+++ app/helpers/application_helper.rb
@@ -20,8 +20,14 @@ module ApplicationHelper
 
   def login_container_helper
     if current_user
-      (link_to "Profile Settings", edit_user_registration_path, class: 'nav-link') +
-      (link_to "Logout", destroy_user_session_path, method: :delete, class: 'nav-link')
+      if current_page?(edit_user_registration_path)
+        (link_to "My Ads", '/', class: 'nav-link') +
+        (link_to "My Profile", '/', class: 'nav-link') +
+        (link_to "New Ad", new_ad_path, class: 'nav-link')
+      else
+        (link_to "Profile Settings", edit_user_registration_path, class: 'nav-link') +
+        (link_to "Logout", destroy_user_session_path, method: :delete, class: 'nav-link')
+      end
     else
       (link_to "Sign Up", new_user_registration_path, class: 'nav-link') +
       (link_to "Login", new_user_session_path, class: 'nav-link')
